+++ src/Symfony/Component/Validator/CHANGELOG.md
@@ -4,7 +4,7 @@ CHANGELOG
 6.4
 ---
 
- * Add `*_ONLY_PRIV` and `*_ONLY_RES` versions to `Ip` constraint
+ * Add `*_ONLY_PRIVATE` and `*_ONLY_RESERVED` versions to `Ip` constraint
  * Add `is_valid` function to the `Expression` constraint, its behavior is the same as `ValidatorInterface::validate`
  * Allow single integer for the `versions` option of the `Uuid` constraint
  * Allow single constraint to be passed to the `constraints` option of the `When` constraint