+++ src/Symfony/Component/Validator/CHANGELOG.md
@@ -19,11 +19,15 @@ CHANGELOG
  * Remove `ValidatorBuilder::disableAnnotationMapping()`, use `ValidatorBuilder::disableAttributeMapping()` instead
  * Remove `AnnotationLoader`, use `AttributeLoader` instead
 
+7.1
+---
+
+* Add `*_NO_PUBLIC`, `*_ONLY_PRIVATE` and `*_ONLY_RESERVED` versions to `Ip` constraint
+* Possibility to use all `Ip` constraint versions for `Cidr` constraint
+
 6.4
 ---
 
- * Add `*_NO_PUBLIC`, `*_ONLY_PRIVATE` and `*_ONLY_RESERVED` versions to `Ip` constraint
- * Possibility to use all `Ip` constraint versions for `Cidr` constraint
  * Add `is_valid` function to the `Expression` constraint, its behavior is the same as `ValidatorInterface::validate`
  * Allow single integer for the `versions` option of the `Uuid` constraint
  * Allow single constraint to be passed to the `constraints` option of the `When` constraint